chore(backend): remove deprecated gui package and consolidate service/dto changes
