+++ backend-api/src/main/java/com/licensis/notaire/config/JpaConfig.java
@@ -3,7 +3,7 @@ package com.licensis.notaire.config;
 import jakarta.persistence.EntityManagerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
-import com.licensis.notaire.servicios.AdministradorJpa;
+import com.licensis.notaire.service.AdministradorJpa;
 
 @Configuration
 public class JpaConfig {
