+++ backend-api/src/main/java/com/licensis/notaire/negocio/ControllerNegocio.java
@@ -31,7 +31,7 @@ import com.licensis.notaire.dto.DtoTipoIdentificacion;
 import com.licensis.notaire.dto.DtoTramite;
 import com.licensis.notaire.dto.DtoUsuario;
 import com.licensis.notaire.dto.exceptions.DtoInvalidoException;
-import com.licensis.notaire.gui.ConstantesGui;
+import com.licensis.notaire.audit.ConstantesGui;
 import com.licensis.notaire.jpa.ConceptoJpaController;
 import com.licensis.notaire.jpa.ConstantesPersistencia;
 import com.licensis.notaire.jpa.CopiaJpaController;
@@ -66,8 +66,8 @@ import com.licensis.notaire.jpa.exceptions.IllegalOrphanException;
 import com.licensis.notaire.jpa.exceptions.NonexistentEntityException;
 import com.licensis.notaire.jpa.exceptions.NonexistentJpaException;
 import com.licensis.notaire.jpa.exceptions.PreexistingEntityException;
-import com.licensis.notaire.servicios.AdministradorJpa;
-import com.licensis.notaire.servicios.AdministradorSesion;
+import com.licensis.notaire.service.AdministradorJpa;
+import com.licensis.notaire.service.AdministradorSesion;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
