+++ backend-api/src/main/java/com/licensis/notaire/service/AdministradorJpa.java
@@ -1,4 +1,4 @@
-package com.licensis.notaire.servicios;
+package com.licensis.notaire.service;
 
 import com.licensis.notaire.jpa.ConceptoJpaController;
 import com.licensis.notaire.jpa.CopiaJpaController;
@@ -34,33 +34,14 @@ import java.util.Iterator;
 import jakarta.persistence.EntityManagerFactory;
 import jakarta.persistence.Persistence;
 
-/**
- * Clase que se encarga de instanciar y administrar todos los JPA. Mantiene una
- * lista de todos los
- * JPA instanciados, haciendo que el proceso de carga de los mismos (que
- * inicialmente es muy lento)
- * solo tenga que realizarse una unica vez. Ademas mantiene una lista de todos
- * los JPA, y los
- * metodos necesarios para poder recuperarlos.
- *
- * @author matias
- */
 public class AdministradorJpa {
 
-    /**
-     * Constante que puede ser utilizada para indicar que una entidad no se pudo
-     * persistir, obtener
-     * de la persistencia.
-     */
     public static final int ERROR = -1;
     private static AdministradorJpa instancia = null;
     private static EntityManagerFactory emf = null;
 
     private static Collection<IPersistenciaJpa> milistaJpas = null;
 
-    /**
-     * Permite establecer el EMF desde el contexto de Spring.
-     */
     public static void setEmf(EntityManagerFactory factory) {
         emf = factory;
         if (instancia != null) {
@@ -68,25 +49,12 @@ public class AdministradorJpa {
         }
     }
 
-    /**
-     * Constructor de {@link AdministradorJpa}. Internamente se inicializa y carga
-     * la lista de JPA.
-     */
     private AdministradorJpa() {
-
-        // aqui ocurre todo el proceso de inicializacion de los JPA
         if (emf != null) {
             AdministradorJpa.cargarListaJpas();
         }
-
     }
 
-    /**
-     * Metodo estatico para obtener la instancia actual de AdministradorJpa
-     * (simgleton)
-     *
-     * @return La instancia actual de AdministradorJpa.
-     */
     public static AdministradorJpa getInstancia() {
         if (AdministradorJpa.instancia == null) {
             instancia = new AdministradorJpa();
@@ -94,27 +62,10 @@ public class AdministradorJpa {
         return instancia;
     }
 
-    /**
-     * Retorna la instancia de EntityManagerFactory.
-     *
-     * @return La instancia actual del EMF.
-     */
     public static EntityManagerFactory getEmf() {
         return emf;
     }
 
-    /**
-     * Permite obtener un JPA especificando el nombre de mismo a traves del metodo
-     * del objeto
-     * Class.getName().
-     *
-     * @param nombreClase El nombre del JPA a buscar. El nombre proviene del objeto
-     *                    Class.getName().
-     * @return miJpa Un JPA (IPersistencia) si se encontro coincidencia.
-     * @throws NonexistentJpaException Esta excepcion es lanzada si el JPA indicado
-     *                                 no existe en la
-     *                                 lista de JPA's.
-     */
     public IPersistenciaJpa obtenerJpa(String nombreClase) throws NonexistentJpaException {
         for (Iterator<IPersistenciaJpa> it = milistaJpas.iterator(); it.hasNext();) {
             IPersistenciaJpa iPersistenciaJpa = it.next();
@@ -134,11 +85,6 @@ public class AdministradorJpa {
         AdministradorJpa.milistaJpas = milistaJpas;
     }
 
-    /**
-     * Metodo estatico que carga las instancia de todos los JPA una unica vez y los
-     * guarda en una
-     * lista para poder ser recuperados posteriormente.
-     */
     private static void cargarListaJpas() {
         AdministradorJpa.milistaJpas = new ArrayList<>();
         AdministradorJpa.milistaJpas.add(new ConceptoJpaController(null, emf));
