+++ notaire-shared/src/main/java/com/licensis/notaire/dto/DtoGestionSummary.java
